Struct dns2::Packet

pub struct Packet {
    pub id: i16,
    pub is_query: bool,
    pub kind: QueryKind,
    pub is_authoritative: bool,
    pub truncated: bool,
    pub recursion_desired: bool,
    pub recursion_available: bool,
    pub response_code: ResponseCode,
    pub question: Vec<Question>,
    pub answer: Vec<Record>,
    pub authority: Vec<Record>,
    pub additional: Vec<Record>,
}

A DNS packet.

Fields

id

ID of the packet.

is_query

Set if this is a query.

kind

Kind of the query.

is_authoritative

Set if the answer is authoritative.

truncated

Set if the packet has been truncated.

recursion_desired

Set if recursion is desired for this question.

recursion_available

Set if recursion is available.

response_code

Response code.

question

Questions.

answer

Answers.

authority

Pointers to authorities.

additional

Additional information.

Methods

impl Packet

fn parse(src: &[u8]) -> Result<(usize, Packet), ()>

Parses the bytes in src as a DNS packet.

Return value

Returns the number of bytes read and the packet on success.

fn format(&self, dst: &mut [u8]) -> Result<usize, FormatError>

Formats the packet into the provided buffer.

Return value

Returns the number of bytes written on success.

fn query(id: i16) -> Packet

Creates a new packet that has all header values preset for a query.